tried to get model accuracy another way (was curious) and think our prediction isn't binding back to the dataframe correctly.
More importantly, can you try the partial plots based on the link and new code around 594? I ran the old partial plots for over a day.. so I started this new approach but this package clashes with my version of R. The irony of CHD hold you back is that this may work for you.. ha!